moved config into src/ for nodemon
